requirements.txt
Django==5.2.11
djangorestframework==3.16.1
psycopg2-binary

